+++ src/modules/database/database.providers.ts
@@ -12,6 +12,7 @@ import { MintedNftEvent } from '../ethEventsScraper/domain/mintNftEvent.entity';
 import { DeployCollectionEvent } from '../ethEventsScraper/domain/deploy-collection-event.entity';
 import { MintingCollection } from '../nft/domain/minting-collection.entity';
 import { LoginChallenge } from '../auth/model/login-challenge.entity';
+import { MintingNft } from '../nft/domain/minting-nft.entity';
 
 // TODO: Add db entities here
 const entities = [
@@ -26,6 +27,7 @@ const entities = [
   DeployCollectionEvent,
   MintingCollection,
   LoginChallenge,
+  MintingNft,
 ];
 
 @Injectable()