+++ src/modules/nft/entrypoints/nft.controller.ts
@@ -12,6 +12,7 @@ import {
   ClassSerializerInterceptor,
   Patch,
   Delete,
+  Query,
 } from '@nestjs/common';
 import {
   DeleteSavedNftParams,
@@ -19,12 +20,16 @@ import {
   EditCollectionParams,
   EditMintingCollectionBody,
   EditMintingCollectionParams,
+  EditMintingNftBody,
   EditSavedNftBody,
   GetCollectionParams,
+  GetMyCollectionsParams,
   GetMyNftsResponse,
+  GetNftParams,
   GetNftTokenURIParams,
   GetUserNftsParams,
   GetUserNftsResponse,
+  PatchMintingNftParams,
   PatchSavedNftParams,
   SaveCollectionBody,
   SaveNftBody,
@@ -69,6 +74,16 @@ export class NftController {
     return await this.nftService.editSavedNft(params.id, req.user.sub, classToPlain(body) as any);
   }
 
+  @Patch('/minting-nfts/:id')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiParam({ name: 'id', description: 'The id of the minting NFT', example: 1, required: true })
+  @ApiBearerAuth()
+  @ApiOperation({ summary: 'Edit minting NFT' })
+  async editMintingNft(@Req() req, @Param() params: PatchMintingNftParams, @Body() body: EditMintingNftBody) {
+    return await this.nftService.editMintingNft(params.id, req.user.sub, body);
+  }
+
   // @Post('/collections')
   // @UseGuards(JwtAuthGuard)
   // @ApiTags('nfts')
@@ -91,7 +106,6 @@ export class NftController {
   @Post('/saved-nfts/:id/file')
   @UseGuards(JwtAuthGuard)
   @UseInterceptors(FileInterceptor('file', nftFileMulterOptions()))
-  @UseInterceptors(ClassSerializerInterceptor)
   @ApiTags('nfts')
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Upload image for nft' })
@@ -102,7 +116,7 @@ export class NftController {
     @Param() params: UploadNftMediaFileParams,
     @UploadedFile() file: Express.Multer.File,
   ) {
-    return await this.nftService.uploadMediaFile(params.id, file);
+    return await this.nftService.uploadSavedNftFile(params.id, req.user.sub, file);
   }
 
   @Get('saved-nfts/:id/token-uri')
@@ -113,7 +127,7 @@ export class NftController {
   @ApiParam({ name: 'id', description: 'The id of the nft', required: true, example: 1 })
   @ApiResponse({ status: 200, description: 'The URLs for tokens metadata', type: 'string', isArray: true })
   async getTokenURI(@Param() params: GetNftTokenURIParams) {
-    return await this.nftService.getTokenURI(params.id);
+    return await this.nftService.getSavedNftTokenURI(params.id);
   }
 
   @Post('nfts/token-uri')
@@ -124,8 +138,8 @@ export class NftController {
   @ApiOperation({ summary: 'Generate the token URI for an NFT' })
   @ApiConsumes('form/multi-part')
   @ApiResponse({ status: 200, description: 'The URLs for tokens metadata', type: 'string', isArray: true })
-  async getNftTokenURI(@Req() req: Request, @UploadedFile() file: Express.Multer.File) {
-    return await this.nftService.getNftTokenURI(req.body, file);
+  async getNftTokenURI(@Req() req: any, @UploadedFile() file: Express.Multer.File) {
+    return await this.nftService.getNftTokenURI(req.user.sub, req.body, file);
   }
 
   @Post('nfts/minting-collections')
@@ -234,8 +248,11 @@ export class NftController {
   @ApiTags('nfts')
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Get the list of all my collections' })
-  async getMyCollections(@Req() req) {
-    return await this.nftService.getMyCollections(req.user.sub);
+  async getMyCollections(@Req() req, @Query() params: GetMyCollectionsParams) {
+    if (params.mintable === 'true') {
+      return await this.nftService.getMyMintableCollections(req.user.sub);
+    }
+    return await this.nftService.getMyOwnedCollections(req.user.sub);
   }
 
   @Delete('saved-nfts/:id')
@@ -252,5 +269,57 @@ export class NftController {
   @ApiOperation({ summary: 'Get data for Collection page' })
   async getCollectionPage(@Param() params: GetCollectionParams) {
     return this.nftService.getCollectionPage(params.address);
+  }
+
+  @Get('pages/my-nfts')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'My NFTs page' })
+  @ApiBearerAuth()
+  async getMyNftsPage(@Req() req) {
+    return this.nftService.getMyNftsPage(req.user.sub);
+  }
+
+  @Get('pages/my-collections/pending')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'Get data to populate the pending section of My Collections page' })
+  @ApiBearerAuth()
+  async getMyCollectionsPending(@Req() req) {
+    return this.nftService.getMyCollectionsPendingPage(req.user.sub);
+  }
+
+  @Get('pages/my-collections/pending/count')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'Get count of the collections in minting state' })
+  @ApiBearerAuth()
+  async getMyCollectionsPendingCount(@Req() req) {
+    return this.nftService.getMyCollectionsPendingCount(req.user.sub);
+  }
+
+  @Get('pages/my-nfts/pending')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'Pending NFTs for My NFTs page' })
+  @ApiBearerAuth()
+  async getMyNftsPendingPage(@Req() req) {
+    return this.nftService.getMyNftsPendingPage(req.user.sub);
+  }
+
+  @Get('pages/my-nfts/pending/count')
+  @UseGuards(JwtAuthGuard)
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'Get count of the nfts in minting state' })
+  @ApiBearerAuth()
+  async getMyNftsPendingCount(@Req() req) {
+    return this.nftService.getMyNftsPendingCount(req.user.sub);
+  }
+
+  @Get('pages/nft/:collectionAddress/:tokenId')
+  @ApiTags('nfts')
+  @ApiOperation({ summary: 'Get data for NFT page' })
+  async getNFTPage(@Param() params: GetNftParams) {
+    return this.nftService.getNftPage(params.collectionAddress, params.tokenId);
   }
 }